court-verification/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
geoalchemy2==0.14.4
shapely==2.0.2
python-dotenv==1.0.0
